This is the first time commit for the scripts in the 52 project testing.
The files included are the basic 2d platformer files I made in the Summer
and Fall of 2012.

The most important ones are:
Scripts/Classes/Actors/Player/DefaultPlayer.js
Scripts/Classes/Default and Base Objects/DefaultGameObject.js
Scripts/Classes/Utility/AnimatedImage.js

DefaultGameObject determines how basic interaction is handled for a static
object in space. This means that if you give something the script for this
you can interact with it via collision and so forth, but it doesn't do
anything interested. It is coupled with the AnimatedImage class so that it
can animate in place as well. The collision can be turned off. There can
be a choice between circle or rectangle collision even though it doesn't
support sphere collision. It is all self reliant. All you need to do is
make a game object and just attach it to the object. On play the correct
items are populated using the initial configuration.

DefaultPlayer takes the DefaultObject further by wrapping it and using it
in a manner that makes the player interactable. The player interactions
are plotted out by interaction and are basically an extension of the
default game object, but in a sense that accounts for movement and
interaction.

God I hope this repository doesn't turn to shit.